requirements.txt
streamlit==1.31.0
pandas==2.2.0
Pillow==10.2.0
pytesseract==0.3.10
